requirements.txt
flask==3.0.0
google-genai>=1.0.0
PyPDF2==3.0.1
markdown==3.5.1
Werkzeug==3.0.1
weasyprint>=67.0
